requirements.txt
Flask==3.1.2
numpy==2.3.5
pillow==12.0.0
